Discord Chat Bot
I've made a Discord Chat Bot for playing Dungeons and Dragons and for other stuff. This was a finger exercise for OOP (class and function structures) but I haven't finished it yet. Since I haven't finished it, it has a inventory and health system for only one player. But anyone can use for unlimited amount of player with a little modification on the code.